Add new "initialize" command to bmfs utility
This commit adds a new "initialize" command to the bmfs utility and
updates the README.md document with examples of how to use it.

The initialize command eliminates the need to use "dd" or another similar
program to create a disk image.  It performs the following operations in
one invocation:
1) Creates a new disk image of a specified size.
2) Formats the disk image.
3) Optionally writes MBR, boot loader, and kernel files onto the disk image.